requirements.txt
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.5.2
Flask-WTF==0.14.2
ForgeryPy3==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
peewee==2.10.1
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
